backend/requirements.txt
django
djangorestframework
google-generativeai
langchain
langchain-google-genai
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
whitenoise
xhtml2pdf